Do not add sources that provide content already provided by other sources
This patch tries to fix a loop when using Rygel and UPnP plugin.

Rygel exposes each external media provider application as a UPnP server. Thus,
each source is shown as a different UPnP server.

Problem arises with UPnP plugin. When Rygel exposes a new server, a new
grilo-upnp source is created for that server, and registered in dbus. Rygel
notices it, and creates a server for that new application. But again,
grilo-upnp source notices it too, and creates a new source for it, and so on,
having a loop.

To prevent it, check before if other source is already providing the same
content; if so, then skip the source.