Implement a 'synchronous' browse operation
Instead of implementing a org.freedesktop.DBus.Properties Get/GetAll
asynchronous version, rely in default dbus-glib implementation, which invokes
get_property().

As get_property() is a synchronous function, but grilo's browse is
asynchronous, implement a synchronous version of it. That is, invokes the
asynchronous browse() and then runs mainloop iteratively until browse finishes
and results are ready to be used. Then it just returns.