Do not load more than once a plugin with the same name
While this fix is in fact to prevent a death block with Rygel, it makes sense
for other situations.

When loading a source, check if there is any other already loaded source that
have the same name; if so, then the source is skipped.

This feature can be avoid with an option (--allow-duplicates/-D).