Notify when provider comes and goes.
Clients must be notified when a provider goes away.

The same when a new provider appears.

First part is through a signal, "destroy", which will tell user that provider
have been removed.

Second part is done with an observer, which will emit the "new" signal
everytime a new provider comes up.