Handle introspection message
From now on I'm starting a re-work to cope with new and big changes in V2 spec.

Roughly speaking, it is a mix between V1 and old V2.

As I need more powerful than glib-bindings is able to provide, I'm dealing with
low-level dbus.

In this first commit, I'm handling the Introspectable interface.