Rename from MediaServer2 to MediaServer1
Last non-published version of specification was MediaServer2, and improvement
over MediaServer1.

After several discussions, a new specification was created, which is mostly
MediaServer1 with some new additions.

Instead of naming it MediaServer3, or reusing MediaServer2 name, as the later
was not published and almost no applications are using MediaServer1
specification, its author decided to reuse the name of MediaServer1.

Thus, this commit renames all files and objects to fit with this new name.

So now library is providing MS1Server and MS1Client.