Add a limit to number of children
In spec if user requests Items, ItemCount, Containers or ContainerCount, we
would need to get all children, and then split between items and containers.

As a source can potentially have thousands of elements, it seems that it can
take a huge time to get these values, getting even a dbus timeout in client
side.

Thus, as a protective measure. allow user to specify up to how many children we
would ask plugin. Thus, none of those values will be greater than the limit.